+++ src/choicebench/methods/library/__init__.py
@@ -2,6 +2,7 @@
 
 from choicebench.methods.direct_mcq import DirectMCQRunner
 from choicebench.methods.library.cyclic_logprob import CyclicLogprobRunner
+from choicebench.methods.library.direct_logprob import DirectLogprobRunner
 from choicebench.methods.library.permutation import PermutationRunner
 from choicebench.methods.library.pride import PriDeRunner
 from choicebench.methods.library.shuffled_baseline import ShuffledBaselineRunner
@@ -9,6 +10,7 @@ from choicebench.methods.library.two_stage import TwoStageRunner
 
 __all__ = [
     "CyclicLogprobRunner",
+    "DirectLogprobRunner",
     "DirectMCQRunner",
     "PermutationRunner",
     "PriDeRunner",